Change the test data set.
The previous mini-cube was centered on a bright galaxy with many lines,
which is not the target of Origin, and was often causing headaches to
make the test pass. As the galaxy was extended, to few background pixels
were available to compute the purity.

The new mini-cube is extracted from HDFS v1.24, at position [500:1600,
235:300, 170:250] (see
/muse/HDFS/private/ORIGIN/Extract-HDFS-Subcube-for-test.ipynb for
details)