Correct mask creation for pixel sources
When the number of pixels above the threshold in the correlation map was
too small, the source mask was set to all (or nearly all) the area. Now,
it only consists in the PSF around the position.

Also, the mask in use in the source detection - to correct the plotting
of correlation maps on the edges of the cube - and the minimum photutils
version is set to 0.5.

In verbose mode, a plot for each step of the mask creation is saved
instead of just for the last step.